refactor: centralize talentforge types
