+++ src/types/index.ts
@@ -1,11 +1,4 @@
-export type {
-  User,
-  ResumeVariant,
-  RolePosting,
-  ApplicationRecord,
-  Recruiter,
-  Thread,
-  Message,
-  Offer,
-  OfferComp,
-} from "./talentforge/models";
+export * from "./talentforge/models";
+export * from "./talentforge/job";
+export * from "./talentforge/resume";
+export * from "./talentforge/types";
